Refactor utbot-intellij-python package names
