+++ utbot-intellij-python/src/main/java/org/utbot/intellij/plugin/python/table/UtPyClassItem.java
@@ -1,33 +1,33 @@
-package org.utbot.intellij.plugin.language.python.table;
+package org.utbot.intellij.plugin.python.table;
 
 import com.intellij.icons.AllIcons;
+import com.jetbrains.python.psi.PyClass;
 import com.jetbrains.python.psi.PyElement;
-import com.jetbrains.python.psi.PyFunction;
 
 import javax.swing.*;
 
-public class UtPyFunctionItem implements UtPyTableItem {
-    private final PyFunction pyFunction;
+public class UtPyClassItem implements UtPyTableItem {
+    private final PyClass pyClass;
     private boolean isChecked;
 
-    public UtPyFunctionItem(PyFunction function) {
-        pyFunction = function;
+    public UtPyClassItem(PyClass clazz) {
+        pyClass = clazz;
         isChecked = false;
     }
 
     @Override
     public PyElement getContent() {
-        return pyFunction;
+        return pyClass;
     }
 
     @Override
     public String getIdName() {
-        return pyFunction.getQualifiedName();
+        return pyClass.getQualifiedName();
     }
 
     @Override
     public Icon getIcon() {
-        return AllIcons.Nodes.Function;
+        return AllIcons.Nodes.Class;
     }
 
     @Override
