+++ utbot-ui-commons/build.gradle.kts
@@ -37,7 +37,6 @@ tasks {
     runIde {
         jvmArgs("-Xmx2048m")
         jvmArgs("--add-exports", "java.desktop/sun.awt.windows=ALL-UNNAMED")
-        androidStudioPath?.let { ideDir.set(file(it)) }
     }
 
     patchPluginXml {
